+++ src/cipher/SkipJack.java
@@ -4,7 +4,7 @@ package cipher;
         64bit Cookbook version, 80bit key.
 */
 public class SkipJack {
-    public static long Encrypt( short[] key, long message ){
+    public static long Encrypt( int[] key, long message ){
         long state = message;
 
         int stepcounter = 0;//k starts at 0
@@ -30,7 +30,7 @@ public class SkipJack {
         return state;        
     }
     
-    public static long Decrypt( short[] key, long message ){
+    public static long Decrypt( int[] key, long message ){
         long state = message;
 
         int stepcounter = 31;//k starts at 31
@@ -56,7 +56,7 @@ public class SkipJack {
         return state; 
     }
     
-    private static long roundA( int step, short[] key, long block ){
+    private static long roundA( int step, int[] key, long block ){
         long w_1i, w_2i, w_3i, w_4i;
         long w_1o, w_2o, w_3o, w_4o;
         w_1i = w_2i = w_3i = w_4i = w_1o = w_2o = w_3o = w_4o = 0L;
@@ -74,15 +74,15 @@ public class SkipJack {
        long ret = w_1o << 48 | w_2o << 32 | w_3o << 16 | w_4o;
        return ret;
     }
-    private static long roundB( int step, short[] key, long block ){
+    private static long roundB( int step, int[] key, long block ){
         long w_1i, w_2i, w_3i, w_4i;
         long w_1o, w_2o, w_3o, w_4o;
         w_1i = w_2i = w_3i = w_4i = w_1o = w_2o = w_3o = w_4o = 0;
 
-        w_1i = (int)(block >>> 48);
-        w_2i = (int)((block >> 32) & 0xFFFFL);
-        w_3i = (int)((block >> 16) & 0xFFFFL);
-        w_4i = (int)(block & 0xFFFFL);
+        w_1i = (block >>> 48);
+        w_2i = ((block >> 32) & 0xFFFFL);
+        w_3i = ((block >> 16) & 0xFFFFL);
+        w_4i = (block & 0xFFFFL);
 
         w_1o = w_4i; 
         w_2o = G( step, key, w_1i );
@@ -93,7 +93,7 @@ public class SkipJack {
         return ret;       
     }
 
-    private static long roundAprime( int step, short[] key, long block ){
+    private static long roundAprime( int step, int[] key, long block ){
         long w_1i, w_2i, w_3i, w_4i;
         long w_1o, w_2o, w_3o, w_4o;
         w_1i = w_2i = w_3i = w_4i = w_1o = w_2o = w_3o = w_4o = 0L;
@@ -111,15 +111,15 @@ public class SkipJack {
        long ret = w_1o << 48 | w_2o << 32 | w_3o << 16 | w_4o;
        return ret;
     }
-    private static long roundBprime( int step, short[] key, long block ){
+    private static long roundBprime( int step, int[] key, long block ){
         long w_1i, w_2i, w_3i, w_4i;
         long w_1o, w_2o, w_3o, w_4o;
         w_1i = w_2i = w_3i = w_4i = w_1o = w_2o = w_3o = w_4o = 0;
 
-        w_1i = (int)(block >>> 48);
-        w_2i = (int)((block >> 32) & 0xFFFFL);
-        w_3i = (int)((block >> 16) & 0xFFFFL);
-        w_4i = (int)(block & 0xFFFFL);
+        w_1i = (block >>> 48);
+        w_2i = ((block >> 32) & 0xFFFFL);
+        w_3i = ((block >> 16) & 0xFFFFL);
+        w_4i = (block & 0xFFFFL);
 
         w_1o = Gprime( step, key, w_2i ); 
         w_2o = Gprime( step, key, w_2i ) ^ w_3i ^ (step + 1);
@@ -130,47 +130,47 @@ public class SkipJack {
         return ret;       
     }
     
-    private static long G( int step, short[] key, long w ){
-        short g_1, g_2, g_3, g_4, g_5, g_6, cv0, cv1, cv2, cv3;
+    private static long G( int step, int[] key, long w ){
+        int g_1, g_2, g_3, g_4, g_5, g_6, cv0, cv1, cv2, cv3;
         g_1 = g_2 = g_3 = g_4 = g_5 = g_6 = cv0 = cv1 = cv2 = cv3 = 0;
 
-        g_1=(short) (w >>> 8);
-        g_2=(short) (w & 0xFF);
+        g_1=(int) (w >>> 8);
+        g_2=(int) (w & 0xFF);
     
         cv0=key[(step * 4) % 10];
         cv1=key[(step * 4 + 1) % 10];
         cv2=key[(step * 4 + 2) % 10];
         cv3=key[(step * 4 + 3) % 10];
 
-        g_3 = (short) (F[g_2 ^ cv0] ^ g_1);
-        g_4 = (short) (F[g_3 ^ cv1] ^ g_2);
-        g_5 = (short) (F[g_4 ^ cv2] ^ g_3);
-        g_6 = (short) (F[g_5 ^ cv3] ^ g_4);
+        g_3 = (F[g_2 ^ cv0] ^ g_1);
+        g_4 = (F[g_3 ^ cv1] ^ g_2);
+        g_5 = (F[g_4 ^ cv2] ^ g_3);
+        g_6 = (F[g_5 ^ cv3] ^ g_4);
         
         return ((long)g_5 << 8) | g_6;
     }  
-    private static long Gprime( int step, short[] key, long w ){
-        short g_1, g_2, g_3, g_4, g_5, g_6, cv0, cv1, cv2, cv3;
+    private static long Gprime( int step, int[] key, long w ){
+        int g_1, g_2, g_3, g_4, g_5, g_6, cv0, cv1, cv2, cv3;
         g_1 = g_2 = g_3 = g_4 = g_5 = g_6 = cv0 = cv1 = cv2 = cv3 = 0;
 
-        g_1=(short) (w & 0xFF);
-        g_2=(short) (w >>> 8);
+        g_1=(int) (w & 0xFF);
+        g_2=(int) (w >>> 8);
         
         cv0=key[(step * 4 + 3) % 10];
         cv1=key[(step * 4 + 2) % 10];
         cv2=key[(step * 4 + 1) % 10];
         cv3=key[(step * 4) % 10];
 
-        g_3 = (short) (F[g_2 ^ cv0] ^ g_1);
-        g_4 = (short) (F[g_3 ^ cv1] ^ g_2);
-        g_5 = (short) (F[g_4 ^ cv2] ^ g_3);
-        g_6 = (short) (F[g_5 ^ cv3] ^ g_4);
+        g_3 = (F[g_2 ^ cv0] ^ g_1);
+        g_4 = (F[g_3 ^ cv1] ^ g_2);
+        g_5 = (F[g_4 ^ cv2] ^ g_3);
+        g_6 = (F[g_5 ^ cv3] ^ g_4);
         
         return ((long)g_6 << 8) | g_5;
     }
 
     // F-Table
-    public static final short[] F = new short[]{
+    public static final int[] F = new int[]{
     	0xa3,0xd7,0x09,0x83,0xf8,0x48,0xf6,0xf4,0xb3,0x21,0x15,0x78,0x99,0xb1,0xaf,0xf9,
     	0xe7,0x2d,0x4d,0x8a,0xce,0x4c,0xca,0x2e,0x52,0x95,0xd9,0x1e,0x4e,0x38,0x44,0x28,
     	0x0a,0xdf,0x02,0xa0,0x17,0xf1,0x60,0x68,0x12,0xb7,0x7a,0xc3,0xe9,0xfa,0x3d,0x53,
